Simplify architectures by removing Broker class
If we need AWS SDK version, we can implement it by Capistrano::NetStorage::S3SDK
